# openapi_schema.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "E-commerce Backend API",
    "description": "A backend API for managing e-commerce products and orders with PostgreSQL persistence",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://your-api-domain.com",
      "description": "Production server"
    }
  ],
  "paths": {
    "/api/products": {
      "get": {
        "summary": "Retrieve all products",
        "description": "Get a list of all existing products in the database",
        "operationId": "getAllProducts",
        "tags": [
          "Products"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                },
                "example": [
                  {
                    "id": 1,
                    "title": "Wireless Headphones",
                    "price": 15000,
                    "image_url": "https://example.com/headphones.jpg"
                  },
                  {
                    "id": 2,
                    "title": "Smartphone Case",
                    "price": 2500,
                    "image_url": "https://example.com/case.jpg"
                  }
                ]
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new product",
        "description": "Add a new product to the database with title, price (in PKR), and image URL",
        "operationId": "createProduct",
        "tags": [
          "Products"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreate"
              },
              "example": {
                "title": "Gaming Keyboard",
                "price": 8500,
                "image_url": "https://example.com/keyboard.jpg"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                },
                "example": {
                  "id": 3,
                  "title": "Gaming Keyboard",
                  "price": 8500,
                  "image_url": "https://example.com/keyboard.jpg"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "summary": "Retrieve a single product",
        "description": "Get a specific product by its unique identifier",
        "operationId": "getProductById",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the product",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                },
                "example": {
                  "id": 1,
                  "title": "Wireless Headphones",
                  "price": 15000,
                  "image_url": "https://example.com/headphones.jpg"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update an existing product",
        "description": "Update specific fields of an existing product. Only provided fields will be modified",
        "operationId": "updateProduct",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the product to update",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdate"
              },
              "example": {
                "title": "Premium Wireless Headphones",
                "price": 18000
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                },
                "example": {
                  "id": 1,
                  "title": "Premium Wireless Headphones",
                  "price": 18000,
                  "image_url": "https://example.com/headphones.jpg"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a product",
        "description": "Remove a product from the database by its unique identifier",
        "operationId": "deleteProduct",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the product to delete",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Product deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders": {
      "post": {
        "summary": "Place a new order",
        "description": "Create a new order with a list of products and their quantities. The API will calculate the total price in PKR",
        "operationId": "createOrder",
        "tags": [
          "Orders"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreate"
              },
              "example": {
                "products": [
                  {
                    "product_id": 1,
                    "quantity": 2
                  },
                  {
                    "product_id": 2,
                    "quantity": 1
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order placed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                },
                "example": {
                  "id": 1,
                  "products": [
                    {
                      "product_id": 1,
                      "title": "Wireless Headphones",
                      "price": 15000,
                      "quantity": 2
                    },
                    {
                      "product_id": 2,
                      "title": "Smartphone Case",
                      "price": 2500,
                      "quantity": 1
                    }
                  ],
                  "total_price": 32500,
                  "created_at": "2025-07-02T10:30:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid input data or product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Retrieve all orders",
        "description": "Get a list of all placed orders with their products, quantities, and total prices in PKR",
        "operationId": "getAllOrders",
        "tags": [
          "Orders"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                },
                "example": [
                  {
                    "id": 1,
                    "products": [
                      {
                        "product_id": 1,
                        "title": "Wireless Headphones",
                        "price": 15000,
                        "quantity": 2
                      },
                      {
                        "product_id": 2,
                        "title": "Smartphone Case",
                        "price": 2500,
                        "quantity": 1
                      }
                    ],
                    "total_price": 32500,
                    "created_at": "2025-07-02T10:30:00Z"
                  }
                ]
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "required": [
          "id",
          "title",
          "price",
          "image_url"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier for the product",
            "example": 1
          },
          "title": {
            "type": "string",
            "description": "Name/title of the product",
            "example": "Wireless Headphones"
          },
          "price": {
            "type": "number",
            "description": "Price of the product in PKR",
            "minimum": 0,
            "example": 15000
          },
          "image_url": {
            "type": "string",
            "format": "uri",
            "description": "URL of the product image",
            "example": "https://example.com/headphones.jpg"
          }
        }
      },
      "ProductCreate": {
        "type": "object",
        "required": [
          "title",
          "price",
          "image_url"
        ],
        "properties": {
          "title": {
            "type": "string",
            "description": "Name/title of the product",
            "minLength": 1,
            "example": "Gaming Keyboard"
          },
          "price": {
            "type": "number",
            "description": "Price of the product in PKR",
            "minimum": 0,
            "example": 8500
          },
          "image_url": {
            "type": "string",
            "format": "uri",
            "description": "URL of the product image",
            "example": "https://example.com/keyboard.jpg"
          }
        }
      },
      "ProductUpdate": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Name/title of the product",
            "minLength": 1,
            "example": "Premium Wireless Headphones"
          },
          "price": {
            "type": "number",
            "description": "Price of the product in PKR",
            "minimum": 0,
            "example": 18000
          },
          "image_url": {
            "type": "string",
            "format": "uri",
            "description": "URL of the product image",
            "example": "https://example.com/premium-headphones.jpg"
          }
        },
        "minProperties": 1
      },
      "OrderCreate": {
        "type": "object",
        "required": [
          "products"
        ],
        "properties": {
          "products": {
            "type": "array",
            "description": "List of products with their quantities",
            "minItems": 1,
            "items": {
              "type": "object",
              "required": [
                "product_id",
                "quantity"
              ],
              "properties": {
                "product_id": {
                  "type": "integer",
                  "description": "ID of the product to order",
                  "minimum": 1,
                  "example": 1
                },
                "quantity": {
                  "type": "integer",
                  "description": "Quantity of the product to order",
                  "minimum": 1,
                  "example": 2
                }
              }
            }
          }
        }
      },
      "Order": {
        "type": "object",
        "required": [
          "id",
          "products",
          "total_price",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier for the order",
            "example": 1
          },
          "products": {
            "type": "array",
            "description": "List of products in the order with their details",
            "items": {
              "type": "object",
              "required": [
                "product_id",
                "title",
                "price",
                "quantity"
              ],
              "properties": {
                "product_id": {
                  "type": "integer",
                  "description": "ID of the product",
                  "example": 1
                },
                "title": {
                  "type": "string",
                  "description": "Name/title of the product",
                  "example": "Wireless Headphones"
                },
                "price": {
                  "type": "number",
                  "description": "Price of the product in PKR",
                  "example": 15000
                },
                "quantity": {
                  "type": "integer",
                  "description": "Quantity of the product in the order",
                  "example": 2
                }
              }
            }
          },
          "total_price": {
            "type": "number",
            "description": "Total price of the order in PKR",
            "example": 32500
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the order was created",
            "example": "2025-07-02T10:30:00Z"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type or code",
            "example": "ValidationError"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "Required field 'title' is missing"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Products",
      "description": "Operations related to product management"
    },
    {
      "name": "Orders",
      "description": "Operations related to order management"
    }
  ]
}